+++ typescript/infra/config/environments/index.ts
@@ -1,9 +1,9 @@
 import { environment as mainnet2 } from './mainnet2';
 import { environment as test } from './test';
-import { environment as testnet3 } from './testnet3';
+import { environment as testnet4 } from './testnet4';
 
 export const environments = {
   test,
-  testnet3,
+  testnet4,
   mainnet2,
 };
