+++ typescript/infra/config/environments/testnet4/core.ts
@@ -7,7 +7,7 @@ import { Contexts } from '../../contexts';
 import { owners } from './owners';
 
 export const core: ChainMap<CoreConfig> = objMap(owners, (local, owner) => {
-  const defaultIsm = aggregationIsm('testnet3', local, Contexts.Hyperlane);
+  const defaultIsm = aggregationIsm('testnet4', local, Contexts.Hyperlane);
   return {
     owner,
     defaultIsm,
