+++ typescript/infra/config/environments/testnet4/core/verification.json
