+++ typescript/infra/config/environments/testnet4/helloworld/hyperlane/verification.json
