+++ typescript/infra/config/environments/testnet4/helloworld/rc/verification.json
