+++ typescript/infra/config/environments/testnet4/infrastructure.ts
@@ -31,14 +31,14 @@ export const infrastructure: InfrastructureConfig = {
       name: 'external-secrets',
       version: '0.5.1',
     },
-    gcpServiceAccountName: 'k8s-external-secrets-testnet3',
+    gcpServiceAccountName: 'k8s-external-secrets-testnet4',
     accessibleGCPSecretPrefixes: [
       'hyperlane-testnet-',
       'testnet-',
-      'hyperlane-testnet3-',
-      'flowcarbon-testnet3-',
-      'rc-testnet3-',
-      'testnet3-',
+      'hyperlane-testnet4-',
+      'flowcarbon-testnet4-',
+      'rc-testnet4-',
+      'testnet4-',
     ],
   },
 };
