+++ typescript/infra/config/environments/testnet4/middleware/queries/verification.json
