+++ typescript/infra/config/environments/testnet4/testquerysender/verification.json
