+++ typescript/infra/helm/helloworld-kathy/values.yaml
@@ -2,7 +2,7 @@ image:
   repository: gcr.io/hyperlane-labs-dev/hyperlane-monorepo
   tag:
 hyperlane:
-  runEnv: testnet3
+  runEnv: testnet4
   context: hyperlane
   # Only used for fetching secrets, so all chains should be included regardless of being skipped
   chains: []
@@ -10,10 +10,10 @@ hyperlane:
   chainsToSkip: []
   cycleOnce: false
 cronjob:
-  schedule: "0 * * * *"
+  schedule: '0 * * * *'
   successfulJobsHistoryLimit: 3
   failedJobsHistoryLimit: 3
 externalSecrets:
   clusterSecretStore:
-nameOverride: ""
-fullnameOverride: ""
+nameOverride: ''
+fullnameOverride: ''
