+++ typescript/infra/src/config/environment.ts
@@ -65,5 +65,5 @@ export const deployEnvToSdkEnv: Record<
 > = {
   test: 'testnet', // TODO: remove this
   mainnet2: 'mainnet',
-  testnet3: 'testnet',
+  testnet4: 'testnet',
 };
